Refactor: inline-style SCSS 코드로 변경
