+++ components/Header.js
@@ -1,4 +1,4 @@
-import Component from './Component.js';
+import Component from "../core/Component.js";
 
 export class Header extends Component {
   template() {
