+++ components/TodoList.js
@@ -1,18 +1,18 @@
-import Component from './Component.js';
-import { TodoCard } from './TodoCard.js';
+import Component from "../core/Component.js";
+import { TodoCard } from "./TodoCard.js";
 
 export class TodoList extends Component {
   template() {
     const {
       list: { title, todos, editting },
-      idx,
+      idx
     } = this.$props;
     return `
           <div class="todo-title">
             <span class="todo-name">${title}
               <span class="count"><span class="number">${
-                todos.length
-              }</span></span>
+      todos.length
+    }</span></span>
             </span>
                 <span class="add">
                     <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -30,36 +30,36 @@ export class TodoList extends Component {
                 </span>
           </div>
         ${
-          editting
-            ? `<div class="wrapper">
+      editting
+        ? `<div class="wrapper">
             <span class="title"></span><span class="content"></span>
             </div>
             <div class="frame">
               <div class="button-left"><span>취소</span></div>
               <div class="button-right"><span>등록</span></div>
             </div>`
-            : ''
-        }
+        : ""
+    }
     ${todos
       .map(
         (todo, idx) =>
           `<div class="${
-            todo.selected ? 'selected' : 'todo-card'
+            todo.selected ? "selected" : "todo-card"
           }" data-idx="${idx}"></div>`
       )
-      .join('')}`;
+      .join("")}`;
   }
 
   mount() {
     const {
-      list: { todos },
+      list: { todos }
     } = this.$props;
     const card = this.select(`.todo-card[data-idx="0"]`);
     todos.forEach(
       (todo, idx) =>
         new TodoCard(this.select(`.todo-card[data-idx="${idx}"]`), {
           todo,
-          idx,
+          idx
         })
     );
   }
